requirements.txt
Django==5.0.4
celery==5.3.6
redis==5.0.4
requests==2.31.0
openpyxl==3.1.2
cryptography==42.0.5
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==22.0.0
flower==2.0.1
